unet/unetB7/requirements.txt
# Core ML/DL frameworks
tensorflow>=2.10.0,<2.16.0
tensorflow-gpu>=2.10.0,<2.16.0  # Optional: for GPU support

# Image processing and computer vision
opencv-python>=4.5.0
Pillow>=8.0.0
albumentations>=1.3.0

# Data manipulation and analysis
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.60.0

# JSON handling (built-in, but explicit for clarity)
# json - built-in module

# File I/O and path handling
pathlib2>=2.3.0  # For older Python versions

# Optional: Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: TensorBoard for visualization
tensorboard>=2.10.0

# Optional: Model optimization
tensorflow-model-optimization>=0.7.0

# Development and testing (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0